Fix syntax error
Expected argument 'elevp1' to be of type 'int32', got 'int64'.

+++ model/atmosphere/advection/tests/stencil_tests/test_face_val_ppm_stencil_02.py
@@ -64,7 +64,7 @@ class TestFaceValPpmStencil02(StencilTest):
         if hasattr(
             elev, "as_scalar"
         ):  # for backwards compatibility, use `as_scalar` unconditionally once minimum gt4py is >= 1.0.4
-            elev = elev.as_scalar()
+            elev = int32(elev.as_scalar())
         elevp1 = elev + int32(1)
 
         return dict(
